pyproject.toml
[tool.poetry]
name = "dingens"
version = "0.1.0"
description = ""
authors = ["Manuel Stausberg <m.stausberg@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"
Jinja2 = "^3.1.2"
Pillow = "^9.3.0"
stability-sdk = "^0.2.9"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
pytest = "^7.2.0"
pylint = "^2.15.6"
aiofiles = "^22.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
